+++ src/main/java/gov/epa/bencloud/server/database/jooq/data/tables/IncidenceValue.java
@@ -56,7 +56,7 @@ public class IncidenceValue extends TableImpl<IncidenceValueRecord> {
     /**
      * The column <code>data.incidence_value.grid_cell_id</code>.
      */
-    public final TableField<IncidenceValueRecord, Integer> GRID_CELL_ID = createField(DSL.name("grid_cell_id"), SQLDataType.INTEGER, this, "");
+    public final TableField<IncidenceValueRecord, Long> GRID_CELL_ID = createField(DSL.name("grid_cell_id"), SQLDataType.BIGINT, this, "");
 
     /**
      * The column <code>data.incidence_value.grid_col</code>.
@@ -161,7 +161,7 @@ public class IncidenceValue extends TableImpl<IncidenceValueRecord> {
     // -------------------------------------------------------------------------
 
     @Override
-    public Row5<Integer, Integer, Integer, Integer, Double> fieldsRow() {
+    public Row5<Integer, Long, Integer, Integer, Double> fieldsRow() {
         return (Row5) super.fieldsRow();
     }
 }
